result events tidy up and lots of event refactoring
