+++ PYME/IO/h5File.py
@@ -6,8 +6,7 @@ import numpy as np
 import traceback
 from . import h5rFile
 
-EVENTS_DTYPE = np.dtype([('EventDescr', 'S256'), ('EventName', 'S32'), ('Time', '<f8')])
-
+from PYME.IO import events
 #file_cache = {}
 
 #openLock = threading.Lock()
@@ -127,12 +126,8 @@ class H5File(h5rFile.H5RFile):
         if filename in ['metadata.json', 'MetaData']:
             self.updateMetadata(json.loads(data))
         elif filename == 'events.json':
-            events = json.loads(data)
-            
-            events_array = np.empty(len(events), dtype=EVENTS_DTYPE)
-            
-            for j, ev in enumerate(events):
-                events_array['EventName'][j], events_array['EventDescr'][j], events_array['Time'][j] = ev
+            evts = json.loads(data)
+            events_array = events.EventLogger.list_to_array(evts)
                 
             self.addEvents(events_array)
